refactor: move test outside of the source for better separation of concern
