+++ src/node-tools/chunks_over_ws_server.ts
@@ -11,7 +11,7 @@ import { WebSocketServer, WebSocket } from 'ws'
 
 import { WorkerPool } from '../node/NodeWorkerPool.js'
 import { ChunksPolling, parseThreeStub } from '../index.js'
-import { getWorldDemoEnv } from '../tests/configs/world_demo_setup.js'
+import { getWorldDemoEnv } from '../../test/configs/world_demo_setup.js'
 
 const SERVER_PORT = 3000
 const POOL_SIZE = 4
