+++ test/configs/procedural_items.ts
