+++ test/configs/schematics_collections.ts
